+++ LCDproc.h
@@ -28,7 +28,7 @@
 class CLCDproc : public CDisplay
 {
 public:
-  CLCDproc(std::string address, unsigned int port, unsigned int localPort, const std::string& callsign, unsigned int dmrid, bool displayClock, bool utc, bool duplex, bool dimOnIdle);
+  CLCDproc(std::string address, unsigned int port, unsigned short localPort, const std::string& callsign, unsigned int dmrid, bool displayClock, bool utc, bool duplex, bool dimOnIdle);
   virtual ~CLCDproc();
 
   virtual bool open();
@@ -73,7 +73,7 @@ protected:
 private:
 	std::string  m_address;
 	unsigned int m_port;
-	unsigned int m_localPort;
+	unsigned short m_localPort;
 	std::string  m_callsign;
 	unsigned int m_dmrid;
 	bool         m_displayClock;
